some sig/unsig warnings fixes
Ignore-this: 6dd201bde4f416dfac6d765145027188

darcs-hash:20100419044336-54782-191b365efed14b88f958ad97622196fd511cffd8.gz

+++ kernel/common/llcommon/lfsStorage.h
@@ -147,7 +147,7 @@ int lfsGetHeader(void *HeaderBuf, size_t HeaderSize);
 // Returns:
 //     file number - success;
 //     -1 - previous file doesn't exist;
-uint64_t lfsGetPrevFileNumber(int FileNum);
+uint64_t lfsGetPrevFileNumber(uint64_t FileNum);
 
 // Returns number of lfs files (can be used to make decision about truncating)
 // Returns:
